HTML: Compile Scrubbers on boot
This makes it possible to configure their behavior on OTP releases.
